Handle vttablet dying during test startup gracefully.
Timeouts are uninformative, unhelpful, and generally frustrating. This
probably isn't the right way to manage subprocesses in Python, but I'm
not about to fix the entire testsuite.

+++ test/tablet.py
@@ -517,6 +517,8 @@ class Tablet(object):
     while True:
       v = utils.get_vars(port or self.port)
       if v == None:
+        if self.proc.poll() is not None:
+          raise utils.TestError('vttablet died while test waiting for state %s' % expected)
         logging.debug(
             '  vttablet %s not answering at /debug/vars, waiting...',
             self.tablet_alias)
@@ -561,14 +563,17 @@ class Tablet(object):
     logging.debug('killing vttablet: %s', self.tablet_alias)
     if self.proc is not None:
       Tablet.tablets_running -= 1
-      self.proc.terminate()
-      self.proc.wait()
+      if self.proc.poll() is None:
+          self.proc.terminate()
+          self.proc.wait()
       self.proc = None
 
   def wait_for_binlog_server_state(self, expected, timeout=30.0):
     while True:
       v = utils.get_vars(self.port)
       if v == None:
+        if self.proc.poll() is not None:
+          raise utils.TestError('vttablet died while test waiting for state %s' % expected)
         logging.debug('  vttablet not answering at /debug/vars, waiting...')
       else:
         if 'UpdateStreamState' not in v:
@@ -590,6 +595,8 @@ class Tablet(object):
     while True:
       v = utils.get_vars(self.port)
       if v == None:
+        if self.proc.poll() is not None:
+          raise utils.TestError('vttablet died while test waiting for state %s' % expected)
         logging.debug('  vttablet not answering at /debug/vars, waiting...')
       else:
         if 'BinlogPlayerMapSize' not in v:
